mathmate-backend/requirements.txt
fastapi
python-multipart
pillow
pix2tex
uvicorn
google-generativeai
python-dotenv